Add venue Change fetching interval Add debug telegram message

+++ sendTgMsg1.py
@@ -9,6 +9,7 @@ from pathlib import Path
 CHAT_ID_ALL = '-590113207'
 CHAT_ID_HOLI = '-563254817'
 CHAT_ID_WEEK = '-562433679'
+CHAT_ID_DEBUG = '166389413'
 
 column_name = ["Date","Day","Holiday","Time","Venue"]
 
@@ -22,17 +23,20 @@ def sendTgMsg(msg,chat_id):
     requests.post('https://api.telegram.org/bot'+TOKEN +
                   '/sendMessage?chat_id='+chat_id+'&text='+msg)
 
+msg = ''
 diff = ''
+curr_data = ''
+last_data = ''
 
 last_file = Path(path("last_result1.csv"))
 curr_file = Path(path("curr_result1.csv"))
-if last_file.is_file():
+if last_file.is_file() and curr_file.is_file():
     print("last file exists")
 
-    cur_data = pd.read_csv(path("curr_result1.csv"),names=column_name)
+    curr_data = pd.read_csv(path("curr_result1.csv"),names=column_name)
     last_data = pd.read_csv(path("last_result1.csv"),names=column_name)
 
-    diff = pd.merge(cur_data,last_data, how='outer',indicator='Exist').sort_values(['Date','Time'])
+    diff = pd.merge(last_data,curr_data, how='outer',indicator='Exist').sort_values(['Date','Time'])
     diff = diff.loc[diff['Exist'] == 'right_only'].drop(['Exist'],axis=1)
 
     os.remove(path('last_result1.csv'))
@@ -49,6 +53,15 @@ else:
     exit(1)
 
 if not diff.empty:
+    # Send debug msg
+    msg = last_data.drop(["Holiday"],axis=1).to_csv(sep = ',', index = False, header = None)
+    sendTgMsg('LAST:',CHAT_ID_DEBUG)
+    sendTgMsg(msg,CHAT_ID_DEBUG)
+
+    msg = curr_data.drop(["Holiday"],axis=1).to_csv(sep = ',', index = False, header = None)
+    sendTgMsg('CURR:',CHAT_ID_DEBUG)
+    sendTgMsg(msg,CHAT_ID_DEBUG)
+
     # Send all diff
     msg = diff.drop(["Holiday"],axis=1).to_csv(sep = ',', index = False, header = None)
     sendTgMsg(msg,CHAT_ID_ALL)
